refactoring - custom element define
