+++ src/ts/view/player/playerForward.ts
@@ -7,3 +7,5 @@ export default class PlayerForward extends View {
     this.textContent = '⬇️'
   }
 }
+
+customElements.define('racingcar-player-forward', PlayerForward)
